# patient-generated-data/documentation/patient-generated-data.openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "US Core R4",
    "version" : "v1"
  },
...
  "servers" : [
    {
      "url" : "https://sandbox-api.va.gov/services/pgd/v0/r4/",
      "description" : "Sandbox",
      "variables" : { }
    },
    {
      "url" : "https://api.va.gov/services/pgd/v0/r4/",
      "description" : "Production",
      "variables" : { }
    }
  ],
  "security" : [
    {
      "OauthFlowSandbox" : [
        "launch/patient",
        "offline_access",
        "patient/Appointment.read",
        "patient/Observation.read",
        "patient/Observation.write",
        "patient/QuestionnaireResponse.read",
        "patient/QuestionnaireResponse.write",
        "system/Questionnaire.read",
        "system/Questionnaire.write"
      ]
    },
    {
      "OauthFlowProduction" : [
        "launch/patient",
        "offline_access",
        "patient/Appointment.read",
        "patient/Observation.read",
        "patient/Observation.write",
        "patient/QuestionnaireResponse.read",
        "patient/QuestionnaireResponse.write",
        "system/Questionnaire.read",
        "system/Questionnaire.write"
      ]
    }
  ],
  "paths" : {
    "/Appointment" : {
      "get" : {
        "tags" : [
          "Appointment"
        ],
        "summary" : "Appointment Search",
        "description" : "https://www.hl7.org/fhir/R4/appointment.html",
        "operationId" : "appointmentSearch",
        "parameters" : [
          {
            "name" : "patient",
            "in" : "query",
            "description" : "An Integration Control Number (ICN) assigned by the Master Patient Index (MPI) that refers to a patient that is participating in the appointment.",
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "_id",
            "in" : "query",
            "description" : "The logical id of the resource. Once assigned, this value never changes.",
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "identifier",
            "in" : "query",
            "description" : "The logical identifier of the resource. Once assigned, this value never changes.",
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "location",
            "in" : "query",
            "description" : "The location ID of the facility where the appointment takes place.",
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "_lastUpdated",
            "in" : "query",
            "description" : "A date or range of dates (maximum of 2) that describe the date that the appointment was last updated.",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          {
            "name" : "date",
            "in" : "query",
            "description" : "A date or range of dates (maximum of 2) that describe the date of the appointment.",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          {
            "name" : "page",
            "in" : "query",
            "description" : "The page number of the search result.",
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 1
            }
          },
          {
            "name" : "_count",
            "in" : "query",
            "description" : "The number of resources that should be returned in a single page. The maximum count size is 100.",
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 30
            }
          },
          {
            "name" : "_sort",
            "in" : "query",
            "description" : "Parameters by which to sort results.",
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "Record found",
            "content" : {
              "application/json+fhir" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AppointmentBundle"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json+fhir" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OperationOutcome"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json+fhir" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OperationOutcome"
                }
              }
            }
          }
        }
      }
    },
    "/Appointment/{id}" : {
      "get" : {
        "tags" : [
          "Appointment"
        ],
        "summary" : "Appointment Read",
        "description" : "https://www.hl7.org/fhir/R4/appointment.html",
        "operationId" : "appointmentRead",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "description" : "The logical id of the resource. Once assigned, this value never changes.",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "Record found",
            "content" : {
              "application/json+fhir" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Appointment"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json+fhir" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OperationOutcome"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json+fhir" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OperationOutcome"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Appointment" : {
        "required" : [
          "participant",
          "resourceType",
          "status"
        ],
        "type" : "object",
        "properties" : {
          "resourceType" : {
            "type" : "string"
          },
          "id" : {
            "pattern" : "[A-Za-z0-9\\-\\.]{1,64}",
            "type" : "string"
          },
          "meta" : {
            "$ref" : "#/components/schemas/Meta"
          },
          "implicitRules" : {
            "pattern" : "\\S*",
            "type" : "string"
          },
          "language" : {
            "pattern" : "[^\\s]+(\\s[^\\s]+)*",
            "type" : "string"
          },
          "text" : {
            "$ref" : "#/components/schemas/Narrative"
          },
          "contained" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Resource"
            }
          },
          "extension" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Extension"
            }
          },
          "modifierExtension" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Extension"
            }
          },
          "identifier" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Identifier"
            }
          },
          "status" : {
            "type" : "string",
            "enum" : [
              "proposed",
              "pending",
              "booked",
              "arrived",
              "fulfilled",
              "cancelled",
              "noshow",
              "entered-in-error",
              "checked-in",
              "waitlist"
            ]
          },
          "cancelationReason" : {
            "$ref" : "#/components/schemas/CodeableConcept"
          },
          "serviceCategory" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CodeableConcept"
            }
          },
          "serviceType" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CodeableConcept"
            }
          },
          "specialty" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CodeableConcept"
            }
          },
          "appointmentType" : {
            "$ref" : "#/components/schemas/CodeableConcept"
          },
          "reasonCode" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CodeableConcept"
            }
          },
          "reasonReference" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Reference"
            }
          },
          "priority" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          },
          "description" : {
            "pattern" : "[ \\r\\n\\t\\S]+",
            "type" : "string"
          },
          "supportingInformation" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Reference"
            }
          },
          "start" : {
            "pattern" : "([0-9]([0-9]([0-9][1-9]|[1-9]0)|[1-9]00)|[1-9]000)-(0[1-9]|1[0-2])-(0[1-9]|[1-2][0-9]|3[0-1])T([01][0-9]|2[0-3]):[0-5][0-9]:([0-5][0-9]|60)(\\.[0-9]+)?(Z|(\\+|-)((0[0-9]|1[0-3]):[0-5][0-9]|14:00))",
            "type" : "string"
          },
          "end" : {
            "pattern" : "([0-9]([0-9]([0-9][1-9]|[1-9]0)|[1-9]00)|[1-9]000)-(0[1-9]|1[0-2])-(0[1-9]|[1-2][0-9]|3[0-1])T([01][0-9]|2[0-3]):[0-5][0-9]:([0-5][0-9]|60)(\\.[0-9]+)?(Z|(\\+|-)((0[0-9]|1[0-3]):[0-5][0-9]|14:00))",
            "type" : "string"
          },
          "minutesDuration" : {
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32"
          },
          "slot" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Reference"
            }
          },
          "created" : {
            "pattern" : "([0-9]([0-9]([0-9][1-9]|[1-9]0)|[1-9]00)|[1-9]000)(-(0[1-9]|1[0-2])(-(0[1-9]|[1-2][0-9]|3[0-1])(T([01][0-9]|2[0-3]):[0-5][0-9]:([0-5][0-9]|60)(\\.[0-9]+)?(Z|(\\+|-)((0[0-9]|1[0-3]):[0-5][0-9]|14:00)))?)?)?",
            "type" : "string"
          },
          "comment" : {
            "pattern" : "[ \\r\\n\\t\\S]+",
            "type" : "string"
          },
          "patientInstruction" : {
            "pattern" : "[ \\r\\n\\t\\S]+",
            "type" : "string"
          },
          "basedOn" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Reference"
            }
          },
          "participant" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AppointmentParticipant"
            }
          },
          "requestedPeriod" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Period"
            }
          }
        },
        "description" : "https://www.hl7.org/fhir/R4/appointment.html",
        "example" : {
          "resourceType" : "Appointment",
          "id" : "I2-09PJ067B07D553B93JCEB49A0JP12QW9",
          "status" : "cancelled",
          "description" : "Scheduled Visit",
          "start" : "2006-06-30T17:00:00Z",
          "participant" : [
            {
              "type" : [
                {
                  "coding" : [
                    {
                      "system" : "http://hl7.org/fhir/v3/ParticipationType",
                      "code" : "PART",
                      "display" : "Participation"
                    }
                  ],
                  "text" : "Participation"
                }
              ],
              "actor" : {
                "reference" : "https://sandbox-api.va.gov/services/r4/v0/Patient/185601V825290",
                "display" : "VETERAN,JOHN Q"
              },
              "required" : "required",
              "status" : "accepted"
            },
            {
              "type" : [
                {
                  "coding" : [
                    {
                      "system" : "http://hl7.org/fhir/v3/ParticipationType",
                      "code" : "PART",
                      "display" : "Participation"
                    }
                  ],
                  "text" : "Participation"
                }
              ],
              "actor" : {
                "reference" : "https://sandbox-api.va.gov/services/r4/v0/Location/I2-LPMZXCR5691VTFT8P0P0123B90GY34TT",
                "display" : "ZZZTG SURG POD SCHEULLER COBI"
              },
              "required" : "required",
              "status" : "accepted"
            }
          ]
        }
      },
      "AppointmentBundle" : {
        "required" : [
          "resourceType",
          "type"
        ],
        "type" : "object",
        "properties" : {
          "resourceType" : {
            "type" : "string"
          },
          "id" : {
            "pattern" : "[A-Za-z0-9\\-\\.]{1,64}",
            "type" : "string"
          },
          "meta" : {
            "$ref" : "#/components/schemas/Meta"
          },
          "implicitRules" : {
            "pattern" : "\\S*",
            "type" : "string"
          },
          "language" : {
            "pattern" : "[^\\s]+(\\s[^\\s]+)*",
            "type" : "string"
          },
          "identifier" : {
            "$ref" : "#/components/schemas/Identifier"
          },
          "type" : {
            "type" : "string",
            "enum" : [
              "document",
              "message",
              "transaction",
              "transaction-response",
              "batch",
              "batch-response",
              "history",
              "searchset",
              "collection"
            ]
          },
          "timestamp" : {
            "pattern" : "([0-9]([0-9]([0-9][1-9]|[1-9]0)|[1-9]00)|[1-9]000)-(0[1-9]|1[0-2])-(0[1-9]|[1-2][0-9]|3[0-1])T([01][0-9]|2[0-3]):[0-5][0-9]:([0-5][0-9]|60)(\\.[0-9]+)?(Z|(\\+|-)((0[0-9]|1[0-3]):[0-5][0-9]|14:00))",
            "type" : "string"
          },
          "total" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          },
          "link" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BundleLink"
            }
          },
          "entry" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AppointmentEntry"
            }
          },
          "signature" : {
            "$ref" : "#/components/schemas/Signature"
          }
        },
        "example" : {
          "resourceType" : "Bundle",
          "type" : "searchset",
          "total" : 1,
          "link" : [
            {
              "relation" : "self",
              "url" : "https://sandbox-api.va.gov/services/r4/v0/Appointment?patient=185601V825290&page=1&_count=15"
            },
            {
              "relation" : "first",
              "url" : "https://sandbox-api.va.gov/services/r4/v0/Appointment?patient=185601V825290&page=1&_count=15"
            },
            {
              "relation" : "last",
              "url" : "https://sandbox-api.va.gov/services/r4/v0/Appointment?patient=185601V825290&page=1&_count=15"
            }
          ],
          "entry" : [
            {
              "fullUrl" : "https://sandbox-api.va.gov/services/r4/v0/Appointment/I2-09PJ067B07D553B93JCEB49A0JP12QW9",
              "resource" : {
                "resourceType" : "Appointment",
                "id" : "I2-09PJ067B07D553B93JCEB49A0JP12QW9",
                "status" : "cancelled",
                "description" : "Scheduled Visit",
                "start" : "2006-06-30T17:00:00Z",
                "participant" : [
                  {
                    "type" : [
                      {
                        "coding" : [
                          {
                            "system" : "http://hl7.org/fhir/v3/ParticipationType",
                            "code" : "PART",
                            "display" : "Participation"
                          }
                        ],
                        "text" : "Participation"
                      }
                    ],
                    "actor" : {
                      "reference" : "https://sandbox-api.va.gov/services/r4/v0/Patient/185601V825290",
                      "display" : "VETERAN,JOHN Q"
                    },
                    "required" : "required",
                    "status" : "accepted"
                  },
                  {
                    "type" : [
                      {
                        "coding" : [
                          {
                            "system" : "http://hl7.org/fhir/v3/ParticipationType",
                            "code" : "PART",
                            "display" : "Participation"
                          }
                        ],
                        "text" : "Participation"
                      }
                    ],
                    "actor" : {
                      "reference" : "https://sandbox-api.va.gov/services/r4/v0/Location/I2-LPMZXCR5691VTFT8P0P0123B90GY34TT",
                      "display" : "ZZZTG SURG POD SCHEULLER COBI"
                    },
                    "required" : "required",
                    "status" : "accepted"
                  }
                ]
              },
              "search" : {
                "mode" : "match"
              }
            }
          ]
        }
      },
      "AppointmentEntry" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "pattern" : "[A-Za-z0-9\\-\\.]{1,64}",
            "type" : "string"
          },
          "extension" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Extension"
            }
          },
          "modifierExtension" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Extension"
            }
          },
          "link" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BundleLink"
            }
          },
          "fullUrl" : {
            "pattern" : "\\S*",
            "type" : "string"
          },
          "resource" : {
            "$ref" : "#/components/schemas/Appointment"
          },
          "search" : {
            "$ref" : "#/components/schemas/Search"
          },
          "request" : {
            "$ref" : "#/components/schemas/Request"
          },
          "response" : {
            "$ref" : "#/components/schemas/Response"
          }
        }
      },
...
      "BundleLink" : {
        "required" : [
          "relation",
          "url"
        ],
        "type" : "object",
        "properties" : {
          "id" : {
            "pattern" : "[A-Za-z0-9\\-\\.]{1,64}",
            "type" : "string"
          },
          "extension" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Extension"
            }
          },
          "modifierExtension" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Extension"
            }
          },
          "relation" : {
            "type" : "string",
            "enum" : [
              "about",
              "alternate",
              "appendix",
              "archives",
              "author",
              "blocked-by",
              "bookmark",
              "canonical",
              "chapter",
              "cite-as",
              "collection",
              "contents",
              "convertedFrom",
              "copyright",
              "create-form",
              "current",
              "describedby",
              "describes",
              "disclosure",
              "dns-prefetch",
              "duplicate",
              "edit",
              "edit-form",
              "edit-media",
              "enclosure",
              "first",
              "glossary",
              "help",
              "hosts",
              "hub",
              "icon",
              "index",
              "intervalAfter",
              "intervalBefore",
              "intervalContains",
              "intervalDisjoint",
              "intervalDuring",
              "intervalEquals",
              "intervalFinishedBy",
              "intervalFinishes",
              "intervalln",
              "intervalMeets",
              "intervalMetBy",
              "intervalOverlappedBy",
              "intervalOverlaps",
              "intervalStartedBy",
              "intervalStarts",
              "item",
              "last",
              "latest-version",
              "license",
              "lrdd",
              "memento",
              "monitor",
              "monitor-group",
              "next",
              "next-archive",
              "nofollow",
              "noreferrer",
              "original",
              "payment",
              "pingback",
              "preconnect",
              "predecessor-version",
              "prefetch",
              "preload",
              "prerender",
              "prev",
              "preview",
              "previous",
              "prev-archive",
              "privacy-policy",
              "profile",
              "related",
              "restconf",
              "replies",
              "search",
              "section",
              "self",
              "service",
              "start",
              "stylesheet",
              "subsection",
              "successor-version",
              "tag",
              "terms-of-service",
              "timegate",
              "timemap",
              "type",
              "up",
              "version-history",
              "via",
              "webmention",
              "working-copy",
              "working-copy-of"
            ]
          },
          "url" : {
            "pattern" : "\\S*",
            "type" : "string"
          }
        }
      },
...
      "Identifier" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "pattern" : "[A-Za-z0-9\\-\\.]{1,64}",
            "type" : "string"
          },
          "extension" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Extension"
            }
          },
          "use" : {
            "type" : "string",
            "enum" : [
              "usual",
              "official",
              "temp",
              "secondary",
              "old"
            ]
          },
          "type" : {
            "$ref" : "#/components/schemas/CodeableConcept"
          },
          "system" : {
            "pattern" : "\\S*",
            "type" : "string"
          },
          "value" : {
            "pattern" : "[ \\r\\n\\t\\S]+",
            "type" : "string"
          },
          "period" : {
            "$ref" : "#/components/schemas/Period"
          },
          "assigner" : {
            "$ref" : "#/components/schemas/Reference"
          }
        },
        "description" : "https://www.hl7.org/fhir/R4/datatypes.html#Identifier"
      },
...
      "Meta" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "pattern" : "[A-Za-z0-9\\-\\.]{1,64}",
            "type" : "string"
          },
          "extension" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Extension"
            }
          },
          "versionId" : {
            "pattern" : "[A-Za-z0-9\\-\\.]{1,64}",
            "type" : "string"
          },
          "lastUpdated" : {
            "pattern" : "([0-9]([0-9]([0-9][1-9]|[1-9]0)|[1-9]00)|[1-9]000)-(0[1-9]|1[0-2])-(0[1-9]|[1-2][0-9]|3[0-1])T([01][0-9]|2[0-3]):[0-5][0-9]:([0-5][0-9]|60)(\\.[0-9]+)?(Z|(\\+|-)((0[0-9]|1[0-3]):[0-5][0-9]|14:00))",
            "type" : "string"
          },
          "source" : {
            "pattern" : "\\S*",
            "type" : "string"
          },
          "profile" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "security" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Coding"
            }
          },
          "tag" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Coding"
            }
          }
        },
        "description" : "https://www.hl7.org/fhir/R4/resource.html#meta"
      },
...
      "OperationOutcome" : {
        "required" : [
          "issue",
          "resourceType"
        ],
        "type" : "object",
        "properties" : {
          "resourceType" : {
            "type" : "string"
          },
          "id" : {
            "pattern" : "[A-Za-z0-9\\-\\.]{1,64}",
            "type" : "string"
          },
          "meta" : {
            "$ref" : "#/components/schemas/Meta"
          },
          "implicitRules" : {
            "pattern" : "\\S*",
            "type" : "string"
          },
          "language" : {
            "pattern" : "[^\\s]+(\\s[^\\s]+)*",
            "type" : "string"
          },
          "text" : {
            "$ref" : "#/components/schemas/Narrative"
          },
          "contained" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Resource"
            }
          },
          "modifierExtension" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Extension"
            }
          },
          "extension" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Extension"
            }
          },
          "issue" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Issue"
            }
          }
        },
        "description" : "https://www.hl7.org/fhir/R4/operationoutcome.html",
        "example" : {
          "resourceType" : "OperationOutcome",
          "issue" : [
            {
              "severity" : "error",
              "code" : "request error",
              "details" : {
                "text" : "This request can not be processed"
              }
            }
          ]
        }
      },
...
      "Resource" : {
        "type" : "object",
        "description" : "https://www.hl7.org/fhir/R4/resource.html"
      },
...
      "Signature" : {
        "required" : [
          "type",
          "when",
          "who"
        ],
        "type" : "object",
        "properties" : {
          "id" : {
            "pattern" : "[A-Za-z0-9\\-\\.]{1,64}",
            "type" : "string"
          },
          "extension" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Extension"
            }
          },
          "type" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Coding"
            }
          },
          "when" : {
            "pattern" : "([0-9]([0-9]([0-9][1-9]|[1-9]0)|[1-9]00)|[1-9]000)-(0[1-9]|1[0-2])-(0[1-9]|[1-2][0-9]|3[0-1])T([01][0-9]|2[0-3]):[0-5][0-9]:([0-5][0-9]|60)(\\.[0-9]+)?(Z|(\\+|-)((0[0-9]|1[0-3]):[0-5][0-9]|14:00))",
            "type" : "string"
          },
          "who" : {
            "$ref" : "#/components/schemas/Reference"
          },
          "onBehalfOf" : {
            "$ref" : "#/components/schemas/Reference"
          },
          "targetFormat" : {
            "pattern" : "[^\\s]+(\\s[^\\s]+)*",
            "type" : "string"
          },
          "sigFormat" : {
            "pattern" : "[^\\s]+(\\s[^\\s]+)*",
            "type" : "string"
          },
          "data" : {
            "pattern" : "(\\s*([0-9a-zA-Z\\+\\=]){4}\\s*)+",
            "type" : "string"
          }
        },
        "description" : "http://hl7.org/fhir/R4/datatypes.html#Signature"
      },
...
    },
    "securitySchemes" : {
      "OauthFlowSandbox" : {
        "type" : "oauth2",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "https://sandbox-api.va.gov/oauth2/pgd/v1/authorization",
            "tokenUrl" : "https://sandbox-api.va.gov/oauth2/pgd/v1/token",
            "scopes" : {
              "launch/patient" : "patient launch",
              "offline_access" : "offline access",
              "patient/Appointment.read" : "read appointments",
              "patient/Observation.read" : "read observations",
              "patient/Observation.write" : "write observations",
              "patient/QuestionnaireResponse.read" : "read questionnaire responses",
              "patient/QuestionnaireResponse.write" : "write questionnaire responses",
              "system/Questionnaire.read" : "read questionnaires",
              "system/Questionnaire.write" : "write questionnaires"
            }
          }
        }
      },
      "OauthFlowProduction" : {
        "type" : "oauth2",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "https://api.va.gov/oauth2/pgd/v1/authorization",
            "tokenUrl" : "https://api.va.gov/oauth2/pgd/v1/token",
            "scopes" : {
              "launch/patient" : "patient launch",
              "offline_access" : "offline access",
              "patient/Appointment.read" : "read appointments",
              "patient/Observation.read" : "read observations",
              "patient/Observation.write" : "write observations",
              "patient/QuestionnaireResponse.read" : "read questionnaire responses",
              "patient/QuestionnaireResponse.write" : "write questionnaire responses",
              "system/Questionnaire.read" : "read questionnaires",
              "system/Questionnaire.write" : "write questionnaires"
            }
          }
        }
      }
...
  }
}